+++ js/components/Header/AbsoluteAnchor.js
@@ -1,0 +1,11 @@
+import styled from 'styled-components';
+import constants from '../../constants';
+
+export default styled.a`
+  position: absolute;
+  top: 0;
+  right: 0;
+  padding: 2%;
+  color: ${constants.midGrey};
+  text-decoration: none;
+`;
