+++ js/components/Header/Subtitle.js
@@ -1,0 +1,8 @@
+import styled from 'styled-components';
+import constants from '../../constants';
+
+export default styled.h2`
+  font-weight: 400;
+  color: ${constants.lightBrandColor};
+  margin-top: 0;
+`;
