+++ js/components/Header/Title.js
@@ -1,0 +1,8 @@
+import styled from 'styled-components';
+import constants from '../../constants';
+
+export default styled.h1`
+  font-size: 3em;
+  color: ${constants.brandColor};
+  margin-bottom: 0.25em;
+`;
