+++ js/components/Header/Wrapper.js
@@ -1,0 +1,7 @@
+import styled from 'styled-components';
+
+export default styled.section`
+  width: 100%;
+  text-align: center;
+  margin: 5em 0;
+`;
